Change getSDKInt return type from Int to Long to match generated go

+++ android/src/main/java/com/tailscale/ipn/App.kt
@@ -300,7 +300,7 @@ class App : UninitializedApp(), libtailscale.AppContext, ViewModelStoreOwner {
 
   override fun getOSVersion(): String = Build.VERSION.RELEASE
 
-  override fun getSDKInt(): Int = Build.VERSION.SDK_INT
+  override fun getSDKInt(): Long = Build.VERSION.SDK_INT.toLong()
 
   override fun isChromeOS(): Boolean {
     return packageManager.hasSystemFeature("android.hardware.type.pc")
